Rename threading module to avoid collection Fix reference to support data in HSC
